ui/requirements.txt
streamlit==1.35.0
plotly==5.22.0
requests==2.32.2
pandas==2.2.2
numpy==1.26.4
Pillow==10.3.0
